+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added time filter in filters
 - Constructor schemes: Edge editing
 - Constructor schemes: Step change element z-order